Fixed Mountains of Bugs
- Something that was merged into here from develop did not play nice with the code I was working on
- Tweaked a ton of code to get it working again

+++ app/src/main/java/com/example/gufyguber/FirebaseManager.java
@@ -145,6 +145,12 @@ public class FirebaseManager {
      * @param rideRequest The valid RideRequest object to store using Firebase
      */
     public void storeRideRequest(RideRequest rideRequest) {
+        // Alternative handler when in test mode
+        if (testMode) {
+            Log.e(TAG, "Cannot store ride requests on Firestore in Test Mode");
+            return;
+        }
+
         CollectionReference requestDoc = FirebaseFirestore.getInstance().collection(RIDE_REQUEST_COLLECTION);
 
         Log.d(TAG, String.format("Storing Ride Request for user [%s]", rideRequest.getRiderUID()));
@@ -181,6 +187,13 @@ public class FirebaseManager {
      * @param returnFunction The callback to use once we've finished retrieving a RideRequest
      */
     public void fetchRideRequest(final String riderUID, final ReturnValueListener<RideRequest> returnFunction) {
+        // Alternative handler when in test mode
+        if (testMode) {
+            Log.e(TAG, "Cannot fetch ride requests from Firestore in Test Mode.");
+            returnFunction.returnValue(null);
+            return;
+        }
+
         DocumentReference requestDoc = FirebaseFirestore.getInstance().collection(RIDE_REQUEST_COLLECTION).document(riderUID);
         requestDoc.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
             @Override
@@ -226,7 +239,9 @@ public class FirebaseManager {
      * @return A ListenerRegistration object that should be used to remove the listener when no longer needed
      */
     public ListenerRegistration listenToRideRequest(final String riderUID, final RideRequestListener onChangedListener) {
+        // Alternative handler when in test mode
         if (testMode) {
+            Log.e(TAG, "Cannot listen to Firestore in Test Mode.");
             return null;
         }
 
@@ -254,6 +269,13 @@ public class FirebaseManager {
      * @param riderUID The Rider UID associated with the ride request to be deleted
      */
     public void deleteRideRequest(final String riderUID, final ReturnValueListener<Boolean> returnFunction) {
+        // Alternative handler when in test mode
+        if (testMode) {
+            Log.e(TAG, "Cannot delete ride request from Firestore in Test Mode.");
+            returnFunction.returnValue(Boolean.FALSE);
+            return;
+        }
+
         DocumentReference requestDoc = FirebaseFirestore.getInstance().collection(RIDE_REQUEST_COLLECTION).document(riderUID);
         requestDoc.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
             @Override
@@ -268,6 +290,13 @@ public class FirebaseManager {
      * @return True if the request acceptance succeeded, false otherwise
      */
     public void driverAcceptRideRequest(final String driverUID, final RideRequest request, final ReturnValueListener<Boolean> returnFunction) {
+        // Alternative handler when in test mode
+        if(testMode) {
+            Log.e(TAG, "Cannot accept ride requests on Firestore in Test Mode.");
+            returnFunction.returnValue(Boolean.FALSE);
+            return;
+        }
+
         fetchRideRequest(request.getRiderUID(), new ReturnValueListener<RideRequest>() {
             @Override
             public void returnValue(RideRequest value) {
@@ -310,6 +339,13 @@ public class FirebaseManager {
      * @param returnFunction
      */
     public void riderAcceptDriverOffer(String riderUID, RideRequest request, final ReturnValueListener<Boolean> returnFunction) {
+        // Alternative handler when in test mode
+        if (testMode) {
+            Log.e(TAG, "Cannot accept driver offers on Firestore while in Test Mode.");
+            returnFunction.returnValue(Boolean.FALSE);
+            return;
+        }
+
         fetchRideRequest(riderUID, new ReturnValueListener<RideRequest>() {
             @Override
             public void returnValue(RideRequest value) {
@@ -329,6 +365,12 @@ public class FirebaseManager {
      * @param returnFunction
      */
     public void riderDeclineDriverOffer(String riderUID, RideRequest request, final ReturnValueListener<Boolean> returnFunction) {
+        // Alternative handler when in test mode
+        if (testMode) {
+            Log.e(TAG, "Cannot decline ride offers on Firestore in Test Mode.");
+            returnFunction.returnValue(Boolean.FALSE);
+            return;
+        }
         fetchRideRequest(riderUID, new ReturnValueListener<RideRequest>() {
             @Override
             public void returnValue(RideRequest value) {
@@ -347,6 +389,13 @@ public class FirebaseManager {
      * @param returnFunction The callback to use once we've finished retrieving a Vehicle
      */
     public void fetchRideRequestsWithStatus(final RideRequest.Status status, final ReturnValueListener<ArrayList<RideRequest>> returnFunction) {
+        // Alternative handler when in test mode
+        if(testMode) {
+            Log.e(TAG, "Cannot fetch collections of ride requests from Firestore in Test Mode.");
+            returnFunction.returnValue(null);
+            return;
+        }
+
         FirebaseFirestore.getInstance().collection(RIDE_REQUEST_COLLECTION).whereEqualTo(STATUS_KEY, status.name()).get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
             @Override
             public void onComplete(@NonNull Task<QuerySnapshot> task) {
@@ -371,7 +420,9 @@ public class FirebaseManager {
     }
 
     public ListenerRegistration listenToAllRideRequests(final DriverRideRequestCollectionListener onChangedListener) {
+        // Alternative handler when in test mode
         if (testMode) {
+            Log.e(TAG, "Cannot listen to Firestore in Test Mode.");
             return null;
         }
 
@@ -400,6 +451,12 @@ public class FirebaseManager {
      * @param rider The Rider to store in our cloud Firestore
      */
     public void storeRiderInfo(Rider rider) {
+        // Alternative handler when in test mode
+        if (testMode) {
+            Log.e(TAG, "Cannot store rider info on Firestore in Test Mode.");
+            return;
+        }
+
         CollectionReference usersDoc = FirebaseFirestore.getInstance().collection(USER_COLLECTION);
 
         Log.d(TAG, "createAccount:" + rider.getUID() + " - " + rider.getEmail());
@@ -433,6 +490,12 @@ public class FirebaseManager {
      * @param returnFunction The callback to use once we've finished retrieving a Rider
      */
     public void fetchRiderInfo(final String riderUID, final ReturnValueListener<Rider> returnFunction) {
+        // Alternative handler when in test mode
+        if (testMode) {
+            Log.e(TAG, "Cannot fetch rider info from Firestore in Test Mode.");
+            returnFunction.returnValue(null);
+            return;
+        }
         DocumentReference riderDoc = FirebaseFirestore.getInstance().collection(USER_COLLECTION).document(riderUID);
         riderDoc.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
             @Override
@@ -463,6 +526,12 @@ public class FirebaseManager {
      * @param riderUID The UID associated with the rider record to be deleted
      */
     public void deleteRiderInfo(final String riderUID) {
+        // Alternative handler when in test mode
+        if (testMode) {
+            Log.e(TAG, "Cannot delete rider records from Firestore in Test Mode.");
+            return;
+        }
+
         // TODO: If we delete a rider, what should we delete that's associated with them? Just their personal info?
         DocumentReference riderDoc = FirebaseFirestore.getInstance().collection(USER_COLLECTION).document(riderUID);
         riderDoc.delete();
@@ -473,6 +542,12 @@ public class FirebaseManager {
      * @param driver The Driver to store in our cloud Firestore.
      */
     public void storeDriverInfo(Driver driver) {
+        // Alternative handler when in test mode
+        if (testMode) {
+            Log.e(TAG, "Cannot store driver info on Firestore in Test Mode.");
+            return;
+        }
+
         CollectionReference usersDoc = FirebaseFirestore.getInstance().collection(USER_COLLECTION);
 
         Log.d(TAG, "createAccount:" + driver.getUID() + " - " + driver.getEmail());
@@ -506,6 +581,13 @@ public class FirebaseManager {
      * @param returnFunction The callback to use once we've finished retrieving a driver (and their vehicle)
      */
     public void fetchDriverInfo(final String driverUID, final ReturnValueListener<Driver> returnFunction) {
+        // Alternative handler when in test mode
+        if(testMode) {
+            Log.e(TAG, "Cannot fetch driver info from Firestore in Test Mode.");
+            returnFunction.returnValue(null);
+            return;
+        }
+
         DocumentReference driverDoc = FirebaseFirestore.getInstance().collection(USER_COLLECTION).document(driverUID);
         driverDoc.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
             @Override
@@ -546,6 +628,12 @@ public class FirebaseManager {
      * @param driverUID The UID associated with the driver record to be deleted
      */
     public void deleteDriverInfo(final String driverUID) {
+        // Alternative handler when in test mode
+        if(testMode) {
+            Log.e(TAG, "Cannot delete driver record from Firestore in Test Mode.");
+            return;
+        }
+
         // TODO: If we delete a driver, what should we delete that's associated with them? Just their personal info?
         DocumentReference driverDoc = FirebaseFirestore.getInstance().collection(USER_COLLECTION).document(driverUID);
         driverDoc.delete();
@@ -557,6 +645,12 @@ public class FirebaseManager {
      * @param vehicle The vehicle that is being registered
      */
     public void storeVehicleInfo(String driverUID, Vehicle vehicle) {
+        // Alternative handler when in test mode
+        if (testMode) {
+            Log.e(TAG, "Cannot store vehicle info on Firestore in Test Mode.");
+            return;
+        }
+
         CollectionReference vehicles = FirebaseFirestore.getInstance().collection(VEHICLE_COLLECTION);
         HashMap<String, Object> vehicleData = new HashMap<>();
 
@@ -587,7 +681,14 @@ public class FirebaseManager {
      */
     public void fetchVehicleInfo(final String driverUID, final ReturnValueListener<Vehicle> returnFunction) {
         //TODO: Decide what constitutes a key for getting vehicle info if drivers have > 1 vehicle
-        //TODO: Fetch information about a vehicle
+
+        // Alternative handler when in test mode
+        if (testMode) {
+            Log.e(TAG, "Cannot fetch vehicle record from Firestore while in Test Mode.");
+            returnFunction.returnValue(null);
+            return;
+        }
+
         DocumentReference vehicleDoc = FirebaseFirestore.getInstance().collection(VEHICLE_COLLECTION).document(driverUID);
         vehicleDoc.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
             @Override
@@ -619,6 +720,14 @@ public class FirebaseManager {
      */
     public void isUserDriver(final String userUID, final ReturnValueListener<Boolean> returnFunction) {
         //TODO: Replace with something better
+
+        // Alternative handler when in test mode
+        if (testMode) {
+            Log.e(TAG, "Cannot use Firestore to determine whether or not a user is a driver in Test Mode.");
+            returnFunction.returnValue(Boolean.FALSE);
+            return;
+        }
+
         DocumentReference userDoc = FirebaseFirestore.getInstance().collection(USER_COLLECTION).document(userUID);
         userDoc.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
             @Override
@@ -644,6 +753,12 @@ public class FirebaseManager {
      * @param driverUID The UID associated with the driver who owns the vehicle
      */
     public void deleteVehicleInfo(final String driverUID) {
+        // Alternative handler when in test mode
+        if (testMode) {
+            Log.e(TAG, "Cannot delete a vehicle record from Firestore in Test Mode.");
+            return;
+        }
+
         DocumentReference vehicleDoc = FirebaseFirestore.getInstance().collection(VEHICLE_COLLECTION).document(driverUID);
         vehicleDoc.delete();
     }
@@ -654,6 +769,12 @@ public class FirebaseManager {
      * @param returnFunction Asynchronously returns true if the user exists in our cloud Firestore, false otherwise
      */
     public void checkUser(String UID, final ReturnValueListener<Boolean> returnFunction) {
+        // Alternative handler when in test mode
+        if (testMode) {
+            Log.e(TAG, "Cannot check Firestore for the presence of a user record in Test Mode.");
+            returnFunction.returnValue(Boolean.FALSE);
+            return;
+        }
 
         final DocumentReference userDoc = FirebaseFirestore.getInstance().collection("users").document(UID);
         userDoc.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
